feat: optimize record check
