+++ provider.go
@@ -2,6 +2,9 @@ package alidns
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"strings"
 
 	"github.com/libdns/libdns"
 )
@@ -38,6 +41,13 @@ func (p *Provider) DeleteRecords(ctx context.Context, zone string, recs []libdns
 	var rls []libdns.Record
 	for _, rec := range recs {
 		ar := alidnsRecordWithZone(rec, zone)
+		if ar.RecID == "" {
+			r0, err := p.queryDomainRecord(ctx, ar.Rr, ar.DName, ar.DTyp, ar.DVal)
+			if err != nil {
+				return nil, err
+			}
+			ar.RecID = r0.RecID
+		}
 		_, err := p.delDomainRecord(ctx, ar)
 		if err != nil {
 			return nil, err
@@ -68,9 +78,14 @@ func (p *Provider) SetRecords(ctx context.Context, zone string, recs []libdns.Re
 	for _, rec := range recs {
 		ar := alidnsRecordWithZone(rec, zone)
 		if ar.RecID == "" {
-			ar.RecID, err = p.addDomainRecord(ctx, ar)
+			r0, err := p.queryDomainRecord(ctx, ar.Rr, ar.DName, ar.DTyp)
 			if err != nil {
-				return nil, err
+				ar.RecID, err = p.addDomainRecord(ctx, ar)
+				if err != nil {
+					return nil, err
+				}
+			} else {
+				ar.RecID = r0.RecID
 			}
 		} else {
 			_, err = p.setDomainRecord(ctx, ar)
@@ -81,6 +96,147 @@ func (p *Provider) SetRecords(ctx context.Context, zone string, recs []libdns.Re
 		rls = append(rls, ar.LibdnsRecord())
 	}
 	return rls, nil
+}
+
+func (p *Provider) getClient() error {
+	return p.getClientWithZone("")
+}
+
+func (p *Provider) getClientWithZone(zone string) error {
+	cred := newCredInfo(p.AccKeyID, p.AccKeySecret, p.RegionID)
+	return p.client.getAliClient(cred, zone)
+}
+
+func (p *Provider) addDomainRecord(ctx context.Context, rc aliDomaRecord) (recID string, err error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClientWithZone(rc.DName)
+	if rc.TTL <= 0 {
+		rc.TTL = 600
+	}
+	p.client.aClient.addReqBody("Action", "AddDomainRecord")
+	p.client.aClient.addReqBody("DomainName", rc.DName)
+	p.client.aClient.addReqBody("RR", rc.Rr)
+	p.client.aClient.addReqBody("Type", rc.DTyp)
+	p.client.aClient.addReqBody("Value", rc.DVal)
+	p.client.aClient.addReqBody("TTL", fmt.Sprintf("%d", rc.TTL))
+	rs := aliResult{}
+	err = p.doAPIRequest(ctx, &rs)
+	recID = rs.RecID
+	if err != nil {
+		return "", err
+	}
+	return recID, err
+}
+
+func (p *Provider) delDomainRecord(ctx context.Context, rc aliDomaRecord) (recID string, err error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClient()
+	p.client.aClient.addReqBody("Action", "DeleteDomainRecord")
+	p.client.aClient.addReqBody("RecordId", rc.RecID)
+	rs := aliResult{}
+	err = p.doAPIRequest(ctx, &rs)
+	recID = rs.RecID
+	if err != nil {
+		return "", err
+	}
+	return recID, err
+}
+
+func (p *Provider) setDomainRecord(ctx context.Context, rc aliDomaRecord) (recID string, err error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClientWithZone(rc.DName)
+	if rc.TTL <= 0 {
+		rc.TTL = 600
+	}
+	p.client.aClient.addReqBody("Action", "UpdateDomainRecord")
+	p.client.aClient.addReqBody("RecordId", rc.RecID)
+	p.client.aClient.addReqBody("RR", rc.Rr)
+	p.client.aClient.addReqBody("Type", rc.DTyp)
+	p.client.aClient.addReqBody("Value", rc.DVal)
+	p.client.aClient.addReqBody("TTL", fmt.Sprintf("%d", rc.TTL))
+	rs := aliResult{}
+	err = p.doAPIRequest(ctx, &rs)
+	recID = rs.RecID
+	if err != nil {
+		return "", err
+	}
+	return recID, err
+}
+
+func (p *Provider) getDomainRecord(ctx context.Context, recID string) (aliDomaRecord, error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClient()
+	p.client.aClient.addReqBody("Action", "DescribeDomainRecordInfo")
+	p.client.aClient.addReqBody("RecordId", recID)
+	rs := aliResult{}
+	err := p.doAPIRequest(ctx, &rs)
+	rec := rs.ToDomaRecord()
+	if err != nil {
+		return aliDomaRecord{}, err
+	}
+	return rec, err
+}
+
+func (p *Provider) queryDomainRecords(ctx context.Context, name string) ([]aliDomaRecord, error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClient()
+	p.client.aClient.addReqBody("Action", "DescribeDomainRecords")
+	p.client.aClient.addReqBody("DomainName", strings.Trim(name, "."))
+	rs := aliResult{}
+	err := p.doAPIRequest(ctx, &rs)
+	if err != nil {
+		return []aliDomaRecord{}, err
+	}
+	return rs.DRecords.Record, err
+}
+
+func (p *Provider) queryDomainRecord(ctx context.Context, rr, name string, recType string, recValue ...string) (aliDomaRecord, error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClient()
+	p.client.aClient.addReqBody("Action", "DescribeDomainRecords")
+	p.client.aClient.addReqBody("DomainName", strings.Trim(name, "."))
+	p.client.aClient.addReqBody("RRKeyWord", rr)
+	if recType != "" {
+		p.client.aClient.addReqBody("TypeKeyWord", recType)
+	}
+	if len(recValue) > 0 && recValue[0] != "" {
+		p.client.aClient.addReqBody("ValueKeyWord", recValue[0])
+	}
+	p.client.aClient.addReqBody("SearchMode", "ADVANCED")
+	rs := aliResult{}
+	err := p.doAPIRequest(ctx, &rs)
+	if err != nil {
+		return aliDomaRecord{}, err
+	}
+	if len(rs.DRecords.Record) == 0 {
+		return aliDomaRecord{}, errors.New("the Record Name of the domain not found")
+	}
+	return rs.DRecords.Record[0], err
+}
+
+// REVERSED:queryMainDomain rseserved for absolute names to name,zone
+func (p *Provider) queryMainDomain(ctx context.Context, name string) (string, string, error) {
+	p.client.mutex.Lock()
+	defer p.client.mutex.Unlock()
+	p.getClient()
+	p.client.aClient.addReqBody("Action", "GetMainDomainName")
+	p.client.aClient.addReqBody("InputString", strings.Trim(name, "."))
+	rs := aliResult{}
+	err := p.doAPIRequest(ctx, &rs)
+	if err != nil {
+		return "", "", err
+	}
+	return rs.Rr, rs.DName, err
+}
+
+func (p *Provider) doAPIRequest(ctx context.Context, result interface{}) error {
+	return p.client.doAPIRequest(ctx, "GET", result)
 }
 
 // Interface guards
